chore: abstract distro dependencies into arrays
This commit focuses on maintainability of the InstallPackages function
by abstracting all hardcoded version numbers and dependencies into
maintainable arrays. Regex word boundaries are used to check for the
presence of version numbers in indexes, removing the need to verbosely
hardcode string literals in the business logic.

Shared dependencies (those common across distros) could further ber
consolidated into, e.g., a shared_depends array.

This is the first of a series of changes geared at improving readability
of the InstallPackages function. The conditional logic can be further
streamlined by using switch statements and namerefs and/or nested
functions.

This change also abstracts the upstream driver URL into a constant
variable defined at the top of the script. If the location of the ISO is
expected to change, consider interpolating parts of the URL (e.g.
'stable-virtio').